refactor: 抽提部分组件
