+++ README.md
@@ -39,7 +39,7 @@
 
 1. 在安装之前要确认你的机器上安装了[`node.js`](https://nodejs.org/zh-cn/)。如果没有安装，同学可以到 [`node.js`](https://nodejs.org/zh-cn/) 的官网下载自己电脑的对应的安装包来安装好环境。
 
-   ![官网](README.assets/node_js_official_website.png)
+   ![官网](https://i.loli.net/2020/03/08/9ReO7y26nJIMjGs.png)
 
 2. 检测是否已安装好`node.js`，在命令行中输入
 
@@ -47,7 +47,7 @@
    npm -v
    ```
 
-   ![检查](README.assets/check.png)
+   ![检查](https://i.loli.net/2020/03/08/2XkClD7WH3qIKgE.png)
 
    如果有版本提示，则表明成功，若无可以考虑下要将`node.js`加入环境变量中
 
@@ -83,7 +83,7 @@
 
    如果前面没出差错，等下会自动跳转到浏览器，此页面就是该项目主页面.
 
-   ![主页](README.assets/website_view.png)
+   ![主页](https://i.loli.net/2020/03/08/NfY7oMdj25nJk9S.png)
 
 此外，我强烈推荐你安装 Prettier 这款 VSCode 格式化代码插件。
