+++ package.json
@@ -66,6 +66,7 @@
     "eslint-config-alloy": "^3.6.0",
     "eslint-plugin-flowtype": "^3.13.0",
     "eslint-plugin-react": "^7.19.0",
+    "node-sass": "^4.13.1",
     "prettier": "1.19.1",
     "typescript": "^3.7.5"
   }
