+++ src/components/Footer.tsx
@@ -1,0 +1,15 @@
+import React from 'react';
+
+export default function Footer() {
+  return (
+    <footer style={{ textAlign: 'center', lineHeight: 2, margin: '2rem 0' }}>
+      Copyright © 2008-2020
+      <a href="https://i.scnu.edu.cn/about/" target="_about" style={{ padding: '0 0.25rem' }}>
+        ISCNU
+      </a>
+      . All rights Reserved.
+      <br />
+      华南师范大学网络协会 版权所有
+    </footer>
+  );
+}
