+++ src/components/HelpDoc/index.scss
@@ -1,0 +1,16 @@
+.doc-container {
+  display: flex;
+  flex-direction: row;
+  padding-left: 2em;
+
+  > nav {
+    min-width: 25vw;
+  }
+}
+
+.doc-article {
+  img {
+    max-width: 80%;
+    max-height: 80vh;
+  }
+}
