+++ src/fragments.module.css
@@ -22,23 +22,6 @@
   padding-top: 1.5rem;
 }
 
-.ScreenPage {
-  position: fixed;
-  top: 0;
-  left: 0;
-  z-index: 1;
-  display: flex;
-  flex-direction: column;
-  background: white;
-  width: 100%;
-  height: 100%;
-}
-
-.Article img {
-  max-width: 80%;
-  max-height: 80vh;
-}
-
 .GettingStart label {
   color: white !important;
   font-size: 12pt !important;
