+++ src/react-app-env.d.ts
@@ -1,0 +1,2 @@
+// eslint-disable-next-line spaced-comment
+/// <reference types="react-scripts" />
